+++ src/wifi/model/minstrel-ht-wifi-manager.h
@@ -50,7 +50,6 @@ struct McsGroup
   uint8_t chWidth; ///< channel width (MHz)
   bool isVht; ///< is VHT?
   bool isSupported; ///< is supported?
-
   // To accurately account for TX times, we separate the TX time of the first
   // MPDU in an A-MPDU from the rest of the MPDUs.
   TxTime ratesTxTimeTable; ///< rates transmit time table
@@ -74,28 +73,21 @@ struct HtRateInfo
    * Given a bit rate and a packet length n bytes.
    */
   Time perfectTxTime;
-
   bool supported;               //!< If the rate is supported.
-
-  uint32_t mcsIndex;            //!< The index in the operationalMcsSet of the WifiRemoteStationManager.
-
+  uint8_t mcsIndex;             //!< The index in the operationalMcsSet of the WifiRemoteStationManager.
   uint32_t retryCount;          //!< Retry limit.
   uint32_t adjustedRetryCount;  //!< Adjust the retry limit for this rate.
   uint32_t numRateAttempt;      //!< Number of transmission attempts so far.
   uint32_t numRateSuccess;      //!< Number of successful frames transmitted so far.
   double prob;                  //!< Current probability within last time interval. (# frame success )/(# total frames)
-
   bool retryUpdated;            //!< If number of retries was updated already.
-
   /**
    * Exponential weighted moving average of probability.
    * EWMA calculation:
    * ewma_prob =[prob *(100 - ewma_level) + (ewma_prob_old * ewma_level)]/100
    */
   double ewmaProb;
-
   double ewmsdProb;             //!< Exponential weighted moving standard deviation of probability.
-
   uint32_t prevNumRateAttempt;  //!< Number of transmission attempts with previous rate.
   uint32_t prevNumRateSuccess;  //!< Number of successful frames transmitted with previous rate.
   uint32_t numSamplesSkipped;   //!< Number of times this rate statistics were not updated because no attempts have been made.
@@ -120,13 +112,10 @@ struct GroupInfo
    */
   uint8_t m_col;                  //!< Sample table column.
   uint8_t m_index;                //!< Sample table index.
-
   bool m_supported;               //!< If the rates of this group are supported by the station.
-
   uint32_t m_maxTpRate;           //!< The max throughput rate of this group.
   uint32_t m_maxTpRate2;          //!< The second max throughput rate of this group.
   uint32_t m_maxProbRate;         //!< The highest success probability rate of this group.
-
   HtMinstrelRate m_ratesTable;    //!< Information about rates of this group.
 };
